fn create_normaltx fixed for BN
